requirements.txt
streamlit
face_recognition
opencv-python
pillow
numpy